==> notebooks/0-computer-matching.ipynb <==
import xml.etree.ElementTree as ET
import os

import pandas as pd
from fuzzywuzzy import fuzz 
from fuzzywuzzy import process 
from tqdm.autonotebook import tqdm
from pandarallel import pandarallel

tqdm.pandas()
pandarallel.initialize(progress_bar=True)

### Load Data Brokers

Import CA data brokers list

fn = '../data/data_brokers/ca-data-brokers.csv'
df = pd.read_csv(fn)
df['state'] = 'CA'
ca = df[['Data Broker Name', 'Email Address', 'Website URL', 'Physical Address', 'state']].copy()
ca.rename(inplace=True, columns={
    'Data Broker Name':'name',
    'Email Address':'email',
    'Website URL':'url',
    'Physical Address':'address'
})

Import VT data brokers list

fn = '../data/data_brokers/vt-data-brokers.csv'
df = pd.read_csv(fn)
df['state'] = "VT"

vt = df[['Data Broker Name:','Address:', 'Email Address:', 'Primary Internet Address:', 'state']].copy()
vt.rename(inplace=True, columns={
    'Data Broker Name:':'name',
    'Address:':'address',
    'Email Address:':'email',
    'Primary Internet Address:':'url'
})

Merge the two

brokers = pd.concat([ca, vt])

Save as output

brokers.to_csv('../data/matching_process/brokers.csv', index=False)

### Load Lobbyist Clients

client_list = []

folder = '../data/lobbying/'
for path, dirs, files in os.walk(folder):
    for file in files:
        fullpath = os.path.join(path, file)
        if file.endswith(".xml"):
            with open(fullpath, "rb") as data:
                tree = ET.parse(data)
                root = tree.getroot()
                for filing in root.iter('Filing'):
                    filing_info = filing.attrib
                    for client in filing.iter('Client'):
                        client_info = client.attrib
                        info = {
                            'filing.id': filing_info['ID'],
                            'filing.period': filing_info['Period'],
                            'filing.year': filing_info['Year'],
                            'client.name': client_info['ClientName'],
                            'client.id': client_info['ClientID'],
                            'client.desc': client_info['GeneralDescription'],
                            'client.state': client_info['ClientState'],
                            'client.country': client_info['ClientCountry']
                        }
                        client_list.append(info)

cf = pd.DataFrame(client_list)

Filter for just 2020 filings

clients = cf[cf['filing.year'] == '2020'].copy()

Add bridge to matches

clients['client.name.check'] = clients['client.name'].str.replace(",","").str.replace(".","").str.upper()

Save as output

clients.to_csv('../data/matching_process/clients.csv', index=False)

### Guess Matches

brokers['name.check'] = brokers['name'].str.replace(",","").str.replace(".","").str.upper()

unique_clients = pd.DataFrame()
unique_clients['client.name.check'] = clients['client.name.check'].unique()

choices = list(brokers['name.check'].unique())
choices.extend([
    'EQUIFAX',
    'EXPERIAN',
    'X-MODE',
    'IHS MARITIME & TRADE',
    'ACXIOM',
    'DELOITTE', 
    'PUBLICIS GROUP', 
    'ORACLE',
    'ACCENTURE FEDERAL SERVICES',
    'RELX',
    'ELSEVIER',
    'LIVERAMP',
    'INMAR',
    'EPSILON DATA'])

def guess(client):
    if client in choices:
        return client, 100
    pick, score = process.extractOne(client, choices)
    return pick, score

### First Pass: Data Broker Name List

unique_clients['guess'] = unique_clients['client.name.check'].parallel_apply(guess)

unique_clients[['guess.name', 'guess.confidence']] = unique_clients['guess'].apply(pd.Series)

### Export for Human Double-Checking

describe = unique_clients['guess.confidence'].describe()

guesses = unique_clients[unique_clients['guess.confidence'] > describe['75%']].sort_values(by='guess.confidence', ascending=False)
guesses.to_csv('../data/matching_process/match-guesses.csv', index=False)
