==> notebooks/2-analysis.ipynb <==
import xml.etree.ElementTree as ET
import json
import glob
import re
from pathlib import Path

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

with open('../data/matching_process/filings.json', 'r') as json_file:
    data = json.load(json_file)

df = pd.read_csv('../data/findings/data-broker-filings.csv')

Create a key to match subsidiaries with their parent companies

subsidiaries_df = df[~df['parent_company'].isna()][['parent_company', 'company']].drop_duplicates().copy()
subsidiaries = pd.Series(subsidiaries_df.parent_company.values,index=subsidiaries_df.company).to_dict()

## Spending

period_spending = df.groupby(['company', 'period', 'is_self_filer']).sum().reset_index().copy()

period_spending = period_spending.sort_values(by='is_self_filer', ascending=False).drop_duplicates(subset=['company', 'period'])

period_spending

period_spending['company'] = period_spending['company'].apply(lambda x: x if not x in subsidiaries.keys() else subsidiaries[x])

total_period_spending = period_spending[['company','period','spending']].groupby(['company', 'period']).sum().sort_values(by=['company', 'period']).copy()


ttl = total_period_spending.reset_index().pivot('company', 'period').reset_index()

ttl['2020 Spending'] = ttl.sum(axis=1)

ttl = ttl.sort_values(by='2020 Spending', ascending=False)

Export Spending

ttl.columns = [b if a =="spending" else a for (a, b) in ttl.columns]

ttl.to_csv('../data/findings/period-spending.csv', index=False)

## Issues

d = []
for f in data['Filings']:
    if f['ID'] in df['filing_id'].unique() and f.get('Issues'):
        for issue in f.get('Issues'):
            d.append({
                'filing.id': f['ID'],
                'category': issue['Code'],
                'issue': issue['SpecificIssue']
            })
issues = pd.DataFrame(d)

issues['cleaned'] = issues.issue.str.lower()
issues['hr_bills'] = issues['cleaned'].str.findall('h\.?r\.? ?\d*')
issues['s_bills'] = issues['cleaned'].str.findall('s\. ?\d*')

issues

def find_number(text):
    num = re.findall(r'[0-9]+',text)
    number = "".join(num)
    if len(number) > 0:
        return int(number)
    return np.nan

issues['hr_bills'] = issues['hr_bills'].apply(lambda l: [find_number(x) for x in l])
issues['hr_bills'] = issues['hr_bills'].apply(lambda l: np.nan if len(l) == 0 else l)

issues['s_bills'] = issues['s_bills'].apply(lambda l: [find_number(x) for x in l])
issues['s_bills'] = issues['s_bills'].apply(lambda l: np.nan if len(l) == 0 else l)

issues['hr_bills'] = issues['hr_bills'].apply(lambda l: list(set(l)) if isinstance(l,list) and (len(l) > 0) else np.nan)
issues['s_bills'] = issues['s_bills'].apply(lambda l: list(set(l)) if isinstance(l,list) and (len(l) > 0) else np.nan)

issues['cleaned'] = issues['cleaned'].str.replace('issues related to', '')
issues['cleaned'] = issues['cleaned'].str.replace('issues relating to', '')
issues['cleaned'] = issues['cleaned'].str.replace('issues surrounding', '')

company_issues = issues.merge(df[['filing_id', 'company', 'lobbying_firm', 'period']], left_on='filing.id', right_on='filing_id')
company_issues['company'] = company_issues['company'].apply(lambda x: x if not x in subsidiaries.keys() else subsidiaries[x])

company_issues
